requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.1.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.10
DAWG-Python==0.7.2
Django==4.2.16
djangorestframework==3.15.2
filelock==3.13.1
fsspec==2024.2.0
idna==3.10
Jinja2==3.1.3
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.43.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
murmurhash==1.0.11
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.2
pillow==10.2.0
preshed==3.0.9
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ru_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.8.0/ru_core_news_sm-3.8.0-py3-none-any.whl#sha256=69978d47b43e2c4f329bebdb155e8e9d3861bba1a58ba25551419dae7d7e07fc
shellingham==1.5.4
smart-open==7.1.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.2
srsly==2.5.0
srt==3.5.3
sympy==1.13.1
thinc==8.3.3
tiktoken==0.8.0
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vosk==0.3.45
wasabi==1.1.3
weasel==0.4.1
websockets==14.1
wrapt==1.17.0
